+++ internal/scheme/scheme.go
@@ -16,7 +16,7 @@ import (
 )
 
 type client struct {
-	config  config.Config
+	config  *config.Config
 	service Ydb_Scheme_V1.SchemeServiceClient
 }