+++ internal/table/client.go
@@ -61,7 +61,7 @@ func newClient(
 	ctx context.Context,
 	cc grpc.ClientConnInterface,
 	builder SessionBuilder,
-	config config.Config,
+	config *config.Config,
 ) *client {
 	onDone := trace.TableOnInit(config.Trace(), &ctx)
 	if builder == nil {
@@ -100,7 +100,7 @@ type client struct {
 	// It must not be nil.
 	build             SessionBuilder
 	cc                grpc.ClientConnInterface
-	config            config.Config
+	config            *config.Config
 	index             map[Session]sessionInfo
 	createInProgress  int           // KIKIMR-9163: in-create-process counter
 	limit             int           // Upper bound for client size.