+++ table/config/config.go
@@ -3,6 +3,7 @@ package config
 import (
 	"time"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/config"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -16,77 +17,7 @@ const (
 	DefaultIdleKeepAliveThreshold          = 2
 )
 
-type Config interface {
-	// OperationTimeout is the maximum amount of time a YDB server will process
-	// an operation. After timeout exceeds YDB will try to cancel operation and
-	// regardless of the cancellation appropriate error will be returned to
-	// the client.
-	// If OperationTimeout is zero then no timeout is used.
-	OperationTimeout() time.Duration
-
-	// OperationCancelAfter is the maximum amount of time a YDB server will process an
-	// operation. After timeout exceeds YDB will try to cancel operation and if
-	// it succeeds appropriate error will be returned to the client; otherwise
-	// processing will be continued.
-	// If OperationCancelAfter is zero then no timeout is used.
-	OperationCancelAfter() time.Duration
-
-	// Trace defines trace over table client calls
-	Trace() trace.Table
-
-	// SizeLimit is an upper bound of pooled sessions.
-	// If SizeLimit is less than or equal to zero then the
-	// DefaultSessionPoolSizeLimit variable is used as a limit.
-	SizeLimit() int
-
-	// KeepAliveMinSize is a lower bound for sessions in the pool. If there are more sessions open, then
-	// the excess idle ones will be closed and removed after IdleKeepAliveThreshold is reached for each of them.
-	// If KeepAliveMinSize is less than zero, then no sessions will be preserved
-	// If KeepAliveMinSize is zero, the DefaultKeepAliveMinSize is used
-	KeepAliveMinSize() int
-
-	// IdleKeepAliveThreshold is a number of keepAlive messages to call before the
-	// session is removed if it is an excess session (see KeepAliveMinSize)
-	// This means that session will be deleted after the expiration of lifetime = IdleThreshold * IdleKeepAliveThreshold
-	// If IdleKeepAliveThreshold is less than zero then it will be treated as infinite and no sessions will
-	// be removed ever.
-	// If IdleKeepAliveThreshold is equal to zero, it will be set to DefaultIdleKeepAliveThreshold
-	IdleKeepAliveThreshold() int
-
-	// IdleLimit is an upper bound of pooled sessions without any activity
-	// within.
-	// IdleLimit int
-
-	// IdleThreshold is a maximum duration between any activity within session.
-	// If this threshold reached, KeepAlive() method will be called on idle
-	// session.
-	//
-	// If IdleThreshold is less than zero then there is no idle limit.
-	// If IdleThreshold is zero, then the DefaultSessionPoolIdleThreshold value
-	// is used.
-	IdleThreshold() time.Duration
-
-	// KeepAliveTimeout limits maximum time spent on KeepAlive request
-	// If KeepAliveTimeout is less than or equal to zero then the
-	// DefaultSessionPoolKeepAliveTimeout is used.
-	KeepAliveTimeout() time.Duration
-
-	// CreateSessionTimeout limits maximum time spent on Create session request
-	// If CreateSessionTimeout is less than or equal to zero then the
-	// DefaultSessionPoolCreateSessionTimeout is used.
-	CreateSessionTimeout() time.Duration
-
-	// DeleteTimeout limits maximum time spent on Delete request
-	// If DeleteTimeout is less than or equal to zero then the
-	// DefaultSessionPoolDeleteTimeout is used.
-	DeleteTimeout() time.Duration
-
-	// PanicCallback returns user-defined panic callback
-	// If nil - panic callback not defined
-	PanicCallback() func(e interface{})
-}
-
-func New(opts ...Option) Config {
+func New(opts ...Option) *Config {
 	c := defaults()
 	for _, o := range opts {
 		o(c)
@@ -94,18 +25,32 @@ func New(opts ...Option) Config {
 	return c
 }
 
-type Option func(*config)
+type Option func(*Config)
 
+// With applies common configuration params
+func With(config config.Common) Option {
+	return func(c *Config) {
+		c.Common = config
+	}
+}
+
+// WithSizeLimit defines upper bound of pooled sessions.
+// If sizeLimit is less than or equal to zero then the
+// DefaultSessionPoolSizeLimit variable is used as a limit.
 func WithSizeLimit(sizeLimit int) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		if sizeLimit > 0 {
 			c.sizeLimit = sizeLimit
 		}
 	}
 }
 
+// WithKeepAliveMinSize defines lower bound for sessions in the pool. If there are more sessions open, then
+// the excess idle ones will be closed and removed after IdleKeepAliveThreshold is reached for each of them.
+// If keepAliveMinSize is less than zero, then no sessions will be preserved
+// If keepAliveMinSize is zero, the DefaultKeepAliveMinSize is used
 func WithKeepAliveMinSize(keepAliveMinSize int) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		if keepAliveMinSize < 0 {
 			keepAliveMinSize = 0
 		}
@@ -113,16 +58,29 @@ func WithKeepAliveMinSize(keepAliveMinSize int) Option {
 	}
 }
 
+// WithIdleKeepAliveThreshold defines number of keepAlive messages to call before the
+// session is removed if it is an excess session (see KeepAliveMinSize)
+// This means that session will be deleted after the expiration of lifetime = IdleThreshold * IdleKeepAliveThreshold
+// If IdleKeepAliveThreshold is less than zero then it will be treated as infinite and no sessions will
+// be removed ever.
+// If IdleKeepAliveThreshold is equal to zero, it will be set to DefaultIdleKeepAliveThreshold
 func WithIdleKeepAliveThreshold(idleKeepAliveThreshold int) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		if idleKeepAliveThreshold > 0 {
 			c.idleKeepAliveThreshold = idleKeepAliveThreshold
 		}
 	}
 }
 
+// WithIdleThreshold sets maximum duration between any activity within session.
+// If this threshold reached, KeepAlive() method will be called on idle
+// session.
+//
+// If idleThreshold is less than zero then there is no idle limit.
+// If idleThreshold is zero, then the DefaultSessionPoolIdleThreshold value
+// is used.
 func WithIdleThreshold(idleThreshold time.Duration) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		if idleThreshold < 0 {
 			idleThreshold = 0
 		}
@@ -130,57 +88,47 @@ func WithIdleThreshold(idleThreshold time.Duration) Option {
 	}
 }
 
+// WithKeepAliveTimeout limits maximum time spent on KeepAlive request
+// If keepAliveTimeout is less than or equal to zero then the DefaultSessionPoolKeepAliveTimeout is used.
 func WithKeepAliveTimeout(keepAliveTimeout time.Duration) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		if keepAliveTimeout > 0 {
 			c.keepAliveTimeout = keepAliveTimeout
 		}
 	}
 }
 
-func WithPanicCallback(cb func(e interface{})) Option {
-	return func(c *config) {
-		c.panicCallback = cb
-	}
-}
-
+// WithCreateSessionTimeout limits maximum time spent on Create session request
+// If createSessionTimeout is less than or equal to zero then the
+// DefaultSessionPoolCreateSessionTimeout is used.
 func WithCreateSessionTimeout(createSessionTimeout time.Duration) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		if createSessionTimeout > 0 {
 			c.createSessionTimeout = createSessionTimeout
 		}
 	}
 }
 
+// WithDeleteTimeout limits maximum time spent on Delete request
+// If deleteTimeout is less than or equal to zero then the DefaultSessionPoolDeleteTimeout is used.
 func WithDeleteTimeout(deleteTimeout time.Duration) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		if deleteTimeout > 0 {
 			c.deleteTimeout = deleteTimeout
 		}
 	}
 }
 
+// WithTrace appends table trace to early defined traces
 func WithTrace(trace trace.Table, opts ...trace.TableComposeOption) Option {
-	return func(c *config) {
+	return func(c *Config) {
 		c.trace = c.trace.Compose(trace, opts...)
 	}
 }
 
-func WithOperationTimeout(operationTimeout time.Duration) Option {
-	return func(c *config) {
-		c.operationTimeout = operationTimeout
-	}
-}
-
-func WithOperationCancelAfter(operationCancelAfter time.Duration) Option {
-	return func(c *config) {
-		c.operationCancelAfter = operationCancelAfter
-	}
-}
-
-type config struct {
-	operationTimeout     time.Duration
-	operationCancelAfter time.Duration
+// Config is a configuration of table client
+type Config struct {
+	config.Common
 
 	sizeLimit              int
 	keepAliveMinSize       int
@@ -191,56 +139,72 @@ type config struct {
 	deleteTimeout          time.Duration
 
 	trace trace.Table
-
-	panicCallback func(e interface{})
 }
 
-func (c *config) PanicCallback() func(e interface{}) {
-	return c.panicCallback
-}
-
-func (c *config) Trace() trace.Table {
+// Trace defines trace over table client calls
+func (c *Config) Trace() trace.Table {
 	return c.trace
 }
 
-func (c *config) OperationTimeout() time.Duration {
-	return c.operationTimeout
-}
-
-func (c *config) OperationCancelAfter() time.Duration {
-	return c.operationCancelAfter
-}
-
-func (c *config) SizeLimit() int {
+// SizeLimit is an upper bound of pooled sessions.
+// If SizeLimit is less than or equal to zero then the
+// DefaultSessionPoolSizeLimit variable is used as a limit.
+func (c *Config) SizeLimit() int {
 	return c.sizeLimit
 }
 
-func (c *config) KeepAliveMinSize() int {
+// KeepAliveMinSize is a lower bound for sessions in the pool. If there are more sessions open, then
+// the excess idle ones will be closed and removed after IdleKeepAliveThreshold is reached for each of them.
+// If KeepAliveMinSize is less than zero, then no sessions will be preserved
+// If KeepAliveMinSize is zero, the DefaultKeepAliveMinSize is used
+func (c *Config) KeepAliveMinSize() int {
 	return c.keepAliveMinSize
 }
 
-func (c *config) IdleKeepAliveThreshold() int {
+// IdleKeepAliveThreshold is a number of keepAlive messages to call before the
+// session is removed if it is an excess session (see KeepAliveMinSize)
+// This means that session will be deleted after the expiration of lifetime = IdleThreshold * IdleKeepAliveThreshold
+// If IdleKeepAliveThreshold is less than zero then it will be treated as infinite and no sessions will
+// be removed ever.
+// If IdleKeepAliveThreshold is equal to zero, it will be set to DefaultIdleKeepAliveThreshold
+func (c *Config) IdleKeepAliveThreshold() int {
 	return c.idleKeepAliveThreshold
 }
 
-func (c *config) IdleThreshold() time.Duration {
+// IdleThreshold is a maximum duration between any activity within session.
+// If this threshold reached, KeepAlive() method will be called on idle
+// session.
+//
+// If IdleThreshold is less than zero then there is no idle limit.
+// If IdleThreshold is zero, then the DefaultSessionPoolIdleThreshold value
+// is used.
+func (c *Config) IdleThreshold() time.Duration {
 	return c.idleThreshold
 }
 
-func (c *config) KeepAliveTimeout() time.Duration {
+// KeepAliveTimeout limits maximum time spent on KeepAlive request
+// If KeepAliveTimeout is less than or equal to zero then the
+// DefaultSessionPoolKeepAliveTimeout is used.
+func (c *Config) KeepAliveTimeout() time.Duration {
 	return c.keepAliveTimeout
 }
 
-func (c *config) CreateSessionTimeout() time.Duration {
+// CreateSessionTimeout limits maximum time spent on Create session request
+// If CreateSessionTimeout is less than or equal to zero then the
+// DefaultSessionPoolCreateSessionTimeout is used.
+func (c *Config) CreateSessionTimeout() time.Duration {
 	return c.createSessionTimeout
 }
 
-func (c *config) DeleteTimeout() time.Duration {
+// DeleteTimeout limits maximum time spent on Delete request
+// If DeleteTimeout is less than or equal to zero then the
+// DefaultSessionPoolDeleteTimeout is used.
+func (c *Config) DeleteTimeout() time.Duration {
 	return c.deleteTimeout
 }
 
-func defaults() *config {
-	return &config{
+func defaults() *Config {
+	return &Config{
 		sizeLimit:              DefaultSessionPoolSizeLimit,
 		keepAliveMinSize:       DefaultKeepAliveMinSize,
 		idleKeepAliveThreshold: DefaultIdleKeepAliveThreshold,